fix: entity폴더 domain으로 바꿈
