+++ ordering/src/main/java/com/careup/ordering/ordering/coupon/domain/CouponStatus.java
@@ -1,0 +1,5 @@
+package com.careup.ordering.ordering.coupon.domain;
+
+public enum CouponStatus {
+    ACTIVE, USED, EXPIRED, CANCELLED
+}
