+++ ordering/src/main/java/com/careup/ordering/ordering/coupon/domain/CouponType.java
@@ -1,0 +1,5 @@
+package com.careup.ordering.ordering.coupon.domain;
+
+public enum CouponType {
+    PERCENT, AMOUNT
+}
