--- ordering/src/main/java/com/careup/ordering/ordering/coupon/entity/CouponStatus.java
@@ -1,5 +1,0 @@
-package com.careup.ordering.ordering.coupon.entity;
-
-public enum CouponStatus {
-    ACTIVE, USED, EXPIRED, CANCELLED
-}
