--- ordering/src/main/java/com/careup/ordering/ordering/coupon/entity/CouponType.java
@@ -1,5 +1,0 @@
-package com.careup.ordering.ordering.coupon.entity;
-
-public enum CouponType {
-    PERCENT, AMOUNT
-}
