+++ ordering/src/main/java/com/careup/ordering/ordering/member/domain/Gender.java
@@ -1,0 +1,6 @@
+package com.careup.ordering.ordering.member.domain;
+
+
+public enum Gender {
+    M, W
+}
