+++ ordering/src/main/java/com/careup/ordering/ordering/order/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package com.careup.ordering.ordering.order.domain;
+
+public enum OrderStatus {
+    PENDING, CONFIRMED, CANCELLED
+}
