+++ ordering/src/main/java/com/careup/ordering/ordering/order/domain/OrderType.java
@@ -1,0 +1,5 @@
+package com.careup.ordering.ordering.order.domain;
+
+public enum OrderType {
+    ONLINE, OFFLINE
+}
