--- ordering/src/main/java/com/careup/ordering/ordering/order/entity/OrderStatus.java
@@ -1,5 +1,0 @@
-package com.careup.ordering.ordering.order.entity;
-
-public enum OrderStatus {
-    PENDING, CONFIRMED, CANCELLED
-}
