+++ ordering/src/main/java/com/careup/ordering/ordering/payment/domain/PaymentStatus.java
@@ -1,0 +1,5 @@
+package com.careup.ordering.ordering.payment.domain;
+
+public enum PaymentStatus {
+    PENDING, COMPLETED, FAILED, CANCELLED
+}
