--- ordering/src/main/java/com/careup/ordering/ordering/payment/entity/PaymentStatus.java
@@ -1,5 +1,0 @@
-package com.careup.ordering.ordering.payment.entity;
-
-public enum PaymentStatus {
-    PENDING, COMPLETED, FAILED, CANCELLED
-}
