+++ ordering/src/main/java/com/careup/ordering/ordering/product/domain/InquiryStatus.java
@@ -1,0 +1,5 @@
+package com.careup.ordering.ordering.product.domain;
+
+public enum InquiryStatus {
+    PENDING, ANSWERED, CLOSED
+}
