+++ ordering/src/main/java/com/careup/ordering/ordering/product/domain/InquiryType.java
@@ -1,0 +1,5 @@
+package com.careup.ordering.ordering.product.domain;
+
+public enum InquiryType {
+    PRODUCT_INFO, DELIVERY, RETURN, ETC
+}
