+++ ordering/src/main/java/com/careup/ordering/ordering/product/domain/Promotion.java
@@ -1,0 +1,53 @@
+package com.careup.ordering.ordering.product.domain;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+import java.time.LocalDate;
+
+@Entity
+@Table(name = "promotion")
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Promotion {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "promotion_id")
+    private Long promotionId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "branch_producs_id", nullable = false)
+    private BranchProduct branchProduct;
+
+    @Column(name = "discount_rate", nullable = false, precision = 5, scale = 2)
+    private BigDecimal discountRate;
+
+    @Column(name = "start_date", nullable = false)
+    private LocalDate startDate;
+
+    @Column(name = "end_date", nullable = false)
+    private LocalDate endDate;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "status", nullable = false)
+    private PromotionStatus status = PromotionStatus.ACTIVE;
+
+    @Builder
+    public Promotion(BranchProduct branchProduct, BigDecimal discountRate,
+                     LocalDate startDate, LocalDate endDate) {
+        this.branchProduct = branchProduct;
+        this.discountRate = discountRate;
+        this.startDate = startDate;
+        this.endDate = endDate;
+    }
+
+    public enum PromotionStatus {
+        ACTIVE, INACTIVE
+    }
+}
