--- ordering/src/main/java/com/careup/ordering/ordering/product/entity/InquiryStatus.java
@@ -1,5 +1,0 @@
-package com.careup.ordering.ordering.product.entity;
-
-public enum InquiryStatus {
-    PENDING, ANSWERED, CLOSED
-}
