--- ordering/src/main/java/com/careup/ordering/ordering/product/entity/InquiryType.java
@@ -1,5 +1,0 @@
-package com.careup.ordering.ordering.product.entity;
-
-public enum InquiryType {
-    PRODUCT_INFO, DELIVERY, RETURN, ETC
-}
